requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.15
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audioop-lts==0.2.1
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
discord.py==2.5.2
docx2txt==0.9
fastapi==0.115.12
frozenlist==1.5.0
h11==0.14.0
idna==3.10
multidict==6.3.1
pdfminer.six==20250327
propcache==0.3.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.1
pydantic_core==2.33.0
python-dotenv==1.1.0
python-multipart==0.0.20
setuptools==75.8.0
sniffio==1.3.1
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.0
uvicorn==0.34.0
wheel==0.45.1
yarl==1.18.3
